Move single LAR validator to persistence project
