+++ api/src/main/scala/hmda/api/HmdaApi.scala
@@ -8,7 +8,7 @@ import akka.stream.ActorMaterializer
 import akka.util.Timeout
 import com.typesafe.config.ConfigFactory
 import hmda.api.http._
-import hmda.api.processing.lar.SingleLarValidation._
+import hmda.persistence.processing.SingleLarValidation._
 import hmda.persistence.institutions.InstitutionPersistence._
 import hmda.persistence.demo.DemoData
 import hmda.api.processing.LocalHmdaEventProcessor._
