+++ api/src/test/scala/hmda/api/http/LarHttpApiSpec.scala
@@ -1,18 +1,19 @@
 package hmda.api.http
 
-import akka.event.{ LoggingAdapter, NoLogging }
+import akka.event.{LoggingAdapter, NoLogging}
 import akka.http.javadsl.model.StatusCodes
-import akka.http.scaladsl.testkit.{ RouteTestTimeout, ScalatestRouteTest }
+import akka.http.scaladsl.testkit.{RouteTestTimeout, ScalatestRouteTest}
 import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.parser.fi.lar.LarCsvParser
-import org.scalatest.{ MustMatchers, WordSpec }
+import org.scalatest.{MustMatchers, WordSpec}
+
 import scala.concurrent.duration._
 import scala.concurrent.ExecutionContext
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
 import akka.util.Timeout
 import hmda.api.RequestHeaderUtils
 import hmda.api.model.ErrorResponse
-import hmda.api.processing.lar.SingleLarValidation
+import hmda.persistence.processing.SingleLarValidation
 import hmda.validation.engine.ValidationError
 import spray.json._
