+++ persistence/src/test/scala/hmda/persistence/processing/SingleLarValidationSpec.scala
@@ -1,16 +1,14 @@
-package hmda.api.processing.lar
+package hmda.persistence.processing
 
 import java.io.File
 
 import akka.testkit.TestProbe
 import hmda.actor.test.ActorSpec
 import hmda.parser.fi.lar.LarCsvParser
-
-import scala.io.Source
-import hmda.api.processing.lar.SingleLarValidation._
-import hmda.validation.context.ValidationContext
+import hmda.persistence.processing.SingleLarValidation._
 
 import scala.concurrent.duration._
+import scala.io.Source
 
 class SingleLarValidationSpec extends ActorSpec {
